Drop active_record/associations directory
