+++ lib/active_record/precount.rb
@@ -1,0 +1,4 @@
+require "active_record"
+require "active_support/lazy_load_hooks"
+
+require "active_record/precount/extend"
