+++ lib/active_record/precount/has_many_extension.rb
@@ -1,4 +1,22 @@
 module ActiveRecord
+  module Associations
+    module Builder
+      class CountLoader < SingularAssociation
+        def self.valid_options
+          [:class, :class_name, :foreign_key]
+        end
+
+        def macro
+          :count_loader
+        end
+
+        def self.valid_dependent_options
+          []
+        end
+      end
+    end
+  end
+
   module Precount
     module Builder
       module HasManyExtension
