+++ lib/activerecord-precount.rb
@@ -1,8 +1,1 @@
-require "active_record"
-require "active_support/lazy_load_hooks"
-
-require "active_record/associations/count_loader"
-require "active_record/associations/builder/count_loader"
-require "active_record/associations/preloader/count_loader"
-
-require "active_record/precount/extend"
+require "active_record/precount"
